Moved TestHarness to common area, putting deprecated GithubAction users into TestHarness to be slowly iterated through and replaced throughout Classy; also updated linting
