+++ packages/portal/backend/src-util/FrontendDatasetGenerator.ts
@@ -1,4 +1,6 @@
 import Log from "../../../common/Log";
+import {Test} from "../../../common/TestHarness";
+
 import {AdminController} from "../src/controllers/AdminController";
 /* istanbul ignore file */
 import {DatabaseController} from "../src/controllers/DatabaseController";
@@ -7,7 +9,6 @@ import {GitHubController} from "../src/controllers/GitHubController";
 import {TeamController} from "../src/controllers/TeamController";
 import {Factory} from "../src/Factory";
 import {PersonKind} from "../src/Types";
-import {Test} from "../test/TestHarness";
 
 /**
  * This sample file shows how to create a bunch of fake data for course testing.
