+++ packages/portal/backend/test/GlobalSpec.ts
@@ -3,9 +3,9 @@ import "mocha";
 
 import Config, {ConfigCourses, ConfigKey} from "../../../common/Config";
 import Log from "../../../common/Log";
+import {Test} from "../../../common/TestHarness";
 
 import {DatabaseController} from "../src/controllers/DatabaseController";
-import {Test} from "./TestHarness";
 
 if (typeof it === 'function') {
     // only if we're running in mocha
