+++ packages/portal/backend/test/controllers/AuthControllerSpec.ts
@@ -1,13 +1,14 @@
 import {expect} from "chai";
 import "mocha";
 
+import {Test} from "../../../../common/TestHarness";
+
 import {AuthController} from "../../src/controllers/AuthController";
 import {DatabaseController} from "../../src/controllers/DatabaseController";
 import {PersonController} from "../../src/controllers/PersonController";
 import {Auth} from "../../src/Types";
 
 import '../GlobalSpec';
-import {Test} from "../TestHarness";
 import './PersonControllerSpec';
 
 describe("AuthController", () => {
