+++ packages/portal/backend/test/controllers/DatabaseControllerSpec.ts
@@ -2,12 +2,12 @@ import {expect} from "chai";
 import "mocha";
 import Config, {ConfigKey} from "../../../../common/Config";
 import Log from "../../../../common/Log";
+import {Test} from "../../../../common/TestHarness";
 
 import {DatabaseController} from "../../src/controllers/DatabaseController";
 import {PersonKind} from "../../src/Types";
 
 import '../GlobalSpec';
-import {Test} from "../TestHarness";
 
 /**
  * This suite seems like a lot of boilerplate, but is crucial to make sure the
