+++ packages/portal/backend/test/controllers/PersonControllerSpec.ts
@@ -1,11 +1,12 @@
 import {expect} from "chai";
 import "mocha";
 
+import {Test} from "../../../../common/TestHarness";
+
 import {PersonController} from "../../src/controllers/PersonController";
 import {Person, PersonKind} from "../../src/Types";
 
 import '../GlobalSpec';
-import {Test} from "../TestHarness";
 import './DeliverablesControllerSpec';
 
 describe("PersonController", () => {
