+++ packages/portal/backend/test/controllers/ResultControllerSpec.ts
@@ -1,12 +1,12 @@
 import {expect} from "chai";
 import "mocha";
+import {Test} from "../../../../common/TestHarness";
 import {GradeReport} from "../../../../common/types/ContainerTypes";
 
 import {AutoTestResultTransport} from "../../../../common/types/PortalTypes";
 import {ResultsController} from "../../src/controllers/ResultsController";
 
 import '../GlobalSpec'; // load first
-import {Test} from "../TestHarness";
 import './AdminControllerSpec'; // make sure it runs before github controller (which should be last of the controllers)
 
 describe("ResultController", () => {
