+++ packages/portal/backend/test/controllers/TeamControllerSpec.ts
@@ -2,6 +2,7 @@ import {expect} from "chai";
 import "mocha";
 
 import Log from "../../../../common/Log";
+import {Test} from "../../../../common/TestHarness";
 
 import {DatabaseController} from "../../src/controllers/DatabaseController";
 import {PersonController} from "../../src/controllers/PersonController";
@@ -9,7 +10,6 @@ import {TeamController} from "../../src/controllers/TeamController";
 import {PersonKind} from "../../src/Types";
 
 import '../GlobalSpec';
-import {Test} from "../TestHarness";
 import './PersonControllerSpec';
 
 describe("TeamController", () => {
